// Malic/both/models/image.ts
export interface ImageI {
	_id?: string;
	complete: boolean;
	extension: string;
	name: string;
	progress: number;
	size: number;
	store: string;
	token: string;
	type: string;
	uploadedAt: Date;
	uploading: boolean;
	url: string;
	userId?: string;
}

export interface ThumbI extends ImageI  {
	originalStore?: string;
	originalId?: string;
}